feat: opt geturlfunc
